minor refactor: read-files-funcs -> utils
